+++ api/views/location.py
@@ -1,77 +1,11 @@
 from django.http import HttpResponse, HttpResponseRedirect, HttpResponseNotFound
-from django.template import RequestContext
-from django.shortcuts import render
-from django.contrib.auth.decorators import login_required
-from django.views.generic import View
-from django.utils import timezone
-from django.contrib.auth.models import User
-from django.core.urlresolvers import reverse
-from django.contrib.sites.models import Site
-from django.shortcuts import get_object_or_404
-
 from jwt_auth.mixins import JSONWebTokenAuthMixin
-from jwt_auth.compat import json
-from core.models import *
-from core.serializers import *
-from commands import *
+from django.views.generic import View
 
-from django.views.decorators.csrf import csrf_exempt
-from rest_framework.renderers import JSONRenderer
-from rest_framework.parsers import JSONParser
+from api.utils.http import JSONResponse
+from api.commands.availabilities import *
 
-
-# Create your views here.
-
-class JSONResponse(HttpResponse):
-    """
-    An HttpResponse that renders its content into JSON.
-    """
-    def __init__(self, data, **kwargs):
-        content = JSONRenderer().render(data)
-        kwargs['content_type'] = 'application/json'
-        super(JSONResponse, self).__init__(content, **kwargs)
-
-
-@csrf_exempt
-def availabilities(request):
-    if request.method == 'POST':
-        data = JSONParser().parse(request)
-
-        command = AddAvailabilityChange(request.user, **data)
-        if command.execute():
-            return JSONResponse(command.result().serialize(), status=201)
-        return JSONResponse(command.result().serialize(), status=400)
-    else:
-        return HttpResponseNotFound("404 not found")
-
-
-@csrf_exempt
-def availability_detail(request, availability_id):
-    ''' Retrieve, update or delete an availability.'''
-    try:
-        availability = Availability.objects.get(pk=availability_id)
-    except Availability.DoesNotExist:
-        return HttpResponse(status=404)
-
-    if request.method == 'GET':
-        serializer = AvailabilitySerializer(availability)
-        return JSONResponse(serializer.data)
-
-    elif request.method == 'PUT':
-        data = JSONParser().parse(request)
-        serializer = AvailabilitySerializer(availability, data=data)
-        if serializer.is_valid():
-            serializer.save()
-            return JSONResponse(serializer.data)
-        return JSONResponse(serializer.errors, status=400)
-
-    elif request.method == 'DELETE':
-        availability.delete()
-        return HttpResponse(status=204)
-
-
-def api_test(request):
-    return HttpResponse(200)
+from core.models import Availability
 
 
 class CurrentLocationOccupancies(JSONWebTokenAuthMixin, View):
